LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/start-edge-in-block-layout-direction-expected.txt



FAIL Horizontal LTR. assert_equals: expected 150 but got 120
FAIL Horizontal RTL. assert_equals: expected 130 but got 150
FAIL Vertical-LR LTR. assert_equals: expected 120 but got 135
FAIL Vertical-LR RTL. assert_equals: expected 120 but got 150
FAIL Vertical-RL LTR. assert_equals: expected -120 but got -150
FAIL Vertical-RL RTL. assert_equals: expected -120 but got -150
